chore: various code move to match architecture diagram
