+++ lib/Domain/BTEGui.cpp
