+++ lib/Domain/Gun.cpp
