+++ platformio.ini
@@ -5,7 +5,7 @@ lib_compat_mode=off
 ; these libraries are for AVR and simulation
 ; targets only, so a reset of this property
 ; is needed in the `native` configuration
-lib_deps =
+lib_deps = 
 
 #upload_protocol=custom
 #upload_flags =
@@ -41,7 +41,7 @@ upload_command = avrdude -vv -pm328p -c avrisp -P/dev/ttyS0 $UPLOAD_FLAGS -U fla
   test_ignore = cross/*
   build_flags = -DNATIVE
   debug_test = noarch/test_gun
-  build_src_filter = ${env.src_filter} -<target.cpp> -<gun.cpp>
+  build_src_filter = ${env.src_filter} -<TargetApp.cpp> -<GunApp.cpp>
 
 [env:native_debug]
   lib_deps =  ${env.lib_deps}
@@ -51,21 +51,22 @@ upload_command = avrdude -vv -pm328p -c avrisp -P/dev/ttyS0 $UPLOAD_FLAGS -U fla
   debug_build_flags = -Og -ggdb3 -g3 -DNATIVE -UAVR
   test_ignore = cross/*
   debug_test = noarch/test_gun
-  build_src_filter = ${env.src_filter} -<target.cpp> -<gun.cpp>
+  build_src_filter = ${env.src_filter} -<TargetApp.cpp> -<GunApp.cpp>
 
 [env:cross]
   lib_deps =  ${env.lib_deps} 
   test_framework = unity
   platform = atmelavr
+  framework = arduino
   board = ATmega328P
   build_type = release
   test_ignore = native/*
   build_flags = -O0
-  build_src_filter = ${env.src_filter} -<consoleApp.cpp>
+  build_src_filter = ${env.src_filter} -<ConsoleTargets.cpp>
 
   [env:target]
   lib_deps =  
-    ${env.lib_deps} 
+    ${env.lib_deps}
     SerialCommands
   framework = arduino
   platform = atmelavr
@@ -74,7 +75,9 @@ upload_command = avrdude -vv -pm328p -c avrisp -P/dev/ttyS0 $UPLOAD_FLAGS -U fla
   board_build.f_cpu = 16000000L
   build_type = release
   build_flags = -DAVR
-  build_src_filter = ${env.src_filter} -<gun.cpp> -<consoleApp.cpp>
+  test_ignore =
+    native/*
+  build_src_filter = ${env.src_filter} -<GunApp.cpp> -<ConsoleTargets.cpp>
 
   [env:gun]
   lib_deps =
@@ -82,4 +85,7 @@ upload_command = avrdude -vv -pm328p -c avrisp -P/dev/ttyS0 $UPLOAD_FLAGS -U fla
   board = attiny84
   build_type = release
   build_flags = -DAVR
-  build_src_filter = ${env.src_filter} -<target.cpp> -<consoleApp.cpp>
+  test_ignore = 
+    native/*
+    noarch/*
+  build_src_filter = ${env.src_filter} -<TargetApp.cpp> -<ConsoleTargets.cpp>
